Added secure (SSL) option for embedded Jetty
Added a 'setSecure' method to call before routing to set the embedded
Jetty server to use secure SSL Sockets with keystore of choice

Up-versioned jetty dependency to get SslContextFactory in support of
this.

Altered SparkTestUtils so that is can do secure connections to allow for
testing of new secure server.  Will use supplied test keystore if needs
be, but this
can be overriden by use of JVM SSL parameters, which are respected in
the
test utility.