Add @FunctionalInterface to Spark's functional interfaces in its API
The java annotation @FunctionalInterface is an informal annotation
type,  to indicate that an interface type declaration is intended to be
a functional interface as defined by the Java Language Specification.

A functional interface has exactly one abstract method (not counting
Object.* or Methods with Default Methods).

With this annotation the compiler will throw an error, if this
interface is changed in a way which would breaking existing Lambda
usages etc.