Fixes #610 and implements removing of routes at runtime. Changed manual http methods "GET" to HTTPMethod.GET to simplify and stop any mistakes/typos. Changed HTTPMethod enums to uppercase.
Know issue with removing routes, if you remove a route that either has a after or brfore filter and the route you remove is the last one (excluding filters) then this will cause a 500 error. I will work on a way for this to be found out and solve this in a later PR.

Added test for removal of routes at runtime.

Preformed a code clean up  (organised imports, removed redundant casting)

Deprecated some methods due to teh changes with HTTPMethod enums.